feat: add SimulateSwap

+++ contracts/libraries/uniswap/TickMath.sol
@@ -11,7 +11,7 @@ library TickMath {
 
     // Review : Line Changed
     // Before   int24 internal constant MAX_TICK = -MIN_TICK;
-    uint24 internal constant MAX_TICK = 887272;
+    int24 internal constant MAX_TICK = 887272;
 
     /// @dev The minimum value that can be returned from #getSqrtRatioAtTick. Equivalent to getSqrtRatioAtTick(MIN_TICK)
     uint160 internal constant MIN_SQRT_RATIO = 4295128739;
@@ -28,7 +28,7 @@ library TickMath {
 
         // Review : Line Changed
         // Before   require(absTick <= uint256(MAX_TICK), 'T');
-        require(absTick <= uint256(MAX_TICK), 'T');
+        require(absTick <= uint256(uint24(MAX_TICK)), 'T');
 
         uint256 ratio = absTick & 0x1 != 0 ? 0xfffcb933bd6fad37aa2d162d1a594001 : 0x100000000000000000000000000000000;
         if (absTick & 0x2 != 0) ratio = (ratio * 0xfff97272373d413259a46990580e213a) >> 128;
